Code review changes
