+++ Guppy/src/Model.cpp
@@ -64,15 +64,3 @@ Model::INDEX Model::Base::getMeshOffset(MESH type, uint8_t offset) {
 }
 
 void Model::Base::updateAggregateBoundingBox(Mesh::Base *pMesh) { updateBoundingBox(pMesh->getBoundingBoxMinMax(false)); }
-
-Model::ColorMesh::ColorMesh(Mesh::Handler &handler, Model::CreateInfo *pCreateInfo,
-                            std::shared_ptr<Material::Base> &pMaterial)
-    : Mesh::Color(handler, (std::string)(pCreateInfo->modelPath + "Color Mesh"), pCreateInfo, pMaterial) {
-    status_ = STATUS::PENDING_VERTICES;
-}
-
-Model::TextureMesh::TextureMesh(Mesh::Handler &handler, Model::CreateInfo *pCreateInfo,
-                                std::shared_ptr<Material::Base> &pMaterial)
-    : Mesh::Texture(handler, (std::string)(pCreateInfo->modelPath + "Texture Mesh"), pCreateInfo, pMaterial) {
-    status_ = STATUS::PENDING_VERTICES;
-}
