+++ Guppy/src/Plane.h
@@ -7,6 +7,10 @@
 #include "Face.h"
 #include "Mesh.h"
 
+// clang-format off
+namespace Mesh      { class Handler; }
+// clang-format on
+
 namespace Plane {
 
 constexpr auto DEFAULT_DIMENSION = 1.0f;
@@ -16,13 +20,17 @@ constexpr auto INDEX_SIZE = 6;
 std::vector<Face> make(bool doubleSided = false);
 
 class Color : public Mesh::Color {
-   public:
+    friend class Mesh::Handler;
+
+   protected:
     Color(Mesh::Handler& handler, Mesh::CreateInfo* pCreateInfo, std::shared_ptr<Material::Base>& pMaterial,
           bool doubleSided = false);
 };
 
 class Texture : public Mesh::Texture {
-   public:
+    friend class Mesh::Handler;
+
+   protected:
     Texture(Mesh::Handler& handler, Mesh::CreateInfo* pCreateInfo, std::shared_ptr<Material::Base>& pMaterial,
             bool doubleSided = false);
 };
