+++ Guppy/src/TextureHandler.cpp
@@ -13,50 +13,6 @@
 
 #define STB_FORMAT VK_FORMAT_R8G8B8A8_UNORM
 
-namespace {
-
-// TODO: move to a Sampler class if I need a bunch of these...
-VkDescriptorSetLayoutBinding getDefaultSamplerBindings(uint32_t binding = 0, uint32_t count = 1) {
-    // VkDescriptorSetLayoutBinding&& getDefaultSamplerBindings(
-    // uint32_t binding = static_cast<uint32_t>(DESCRIPTOR::DEFAULT_SAMPLER), uint32_t count = 1) {
-    assert(false);
-    VkDescriptorSetLayoutBinding layoutBinding = {};
-    layoutBinding.binding = binding;
-    layoutBinding.descriptorType = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
-    layoutBinding.descriptorCount = count;
-    layoutBinding.stageFlags = VK_SHADER_STAGE_FRAGMENT_BIT;
-    layoutBinding.pImmutableSamplers = nullptr;  // Optional
-    return layoutBinding;
-}
-
-VkWriteDescriptorSet getDefaultWrite() {
-    assert(false);
-    VkWriteDescriptorSet write = {};
-    write.sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
-    // write.dstBinding = static_cast<uint32_t>(DESCRIPTOR::DEFAULT_SAMPLER);
-    write.dstArrayElement = 0;
-    write.descriptorType = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
-    write.descriptorCount = 1;
-    return write;
-}
-
-VkCopyDescriptorSet getDefaultCopy() {
-    assert(false);
-    /*  1/9/19 I am not using this code because I think that this is only really necessary
-        if you want to reuse the writes but with different bindings... I am currently not
-        to sure about any of this.
-    */
-    VkCopyDescriptorSet copy = {};
-    copy.sType = VK_STRUCTURE_TYPE_COPY_DESCRIPTOR_SET;
-    // copy.srcBinding = static_cast<uint32_t>(DESCRIPTOR::DEFAULT_UNIFORM);
-    copy.srcArrayElement = 0;
-    // copy.dstBinding = static_cast<uint32_t>(DESCRIPTOR::DEFAULT_UNIFORM);
-    copy.dstArrayElement = 0;
-    return copy;
-}
-
-}  // namespace
-
 Texture::Handler::Handler(Game* pGame) : Game::Handler(pGame) {}
 
 void Texture::Handler::init() {
