+++ Guppy/src/Vertex.h
@@ -6,7 +6,6 @@
 #include <glm/glm.hpp>
 #define GLM_ENABLE_EXPERIMENTAL
 #include <glm/gtx/hash.hpp>
-#include <variant>
 #include <vector>
 #include <vulkan/vulkan.h>
 
@@ -29,9 +28,6 @@ struct Texture {
     glm::vec3 tangent;
     glm::vec3 bitangent;
 };
-
-using Variant = std::variant<Vertex::Color, Vertex::Texture>;
-using Vertices = std::variant<std::vector<Vertex::Color>, std::vector<Vertex::Texture>>;
 
 /*  This is used for .obj loading, and the Face class. At some point I might just use
     this instead of breaking the data up because its a pain.
